requirements.txt
openai==2.32.0
requests==2.33.1
beautifulsoup4==4.14.3
scikit-learn==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
streamlit==1.37.0
playwright